add login interceptor and jwt

+++ communitygroup_user/src/main/java/com/communitygroup/user/controller/AdminController.java
@@ -1,10 +1,12 @@
 package com.communitygroup.user.controller;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import com.communitygroup.user.pojo.Admin;
 import com.communitygroup.user.pojo.User;
 import com.communitygroup.user.service.AdminService;
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -18,6 +20,10 @@ import org.springframework.web.bind.annotation.RestController;
 import entity.PageResult;
 import entity.Result;
 import entity.StatusCode;
+import util.JwtUtil;
+
+import javax.servlet.http.HttpServletRequest;
+
 /**
  * 控制器层
  * @author Administrator
@@ -30,6 +36,11 @@ public class AdminController {
 
 	@Autowired
 	private AdminService adminService;
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@Autowired
+	HttpServletRequest httpServletRequest;
 	
 	
 	/**
@@ -102,6 +113,10 @@ public class AdminController {
 	 */
 	@RequestMapping(value="/{id}",method= RequestMethod.DELETE)
 	public Result delete(@PathVariable String id ){
+		Claims admin_claim = (Claims) httpServletRequest.getAttribute("admin_claim");
+		if (admin_claim == null){
+			return new Result(false, StatusCode.LOGINERROR, "权限不足");
+		}
 		adminService.deleteById(id);
 		return new Result(true,StatusCode.OK,"删除成功");
 	}
@@ -113,11 +128,16 @@ public class AdminController {
 	 */
 	@RequestMapping(value = "/login", method = RequestMethod.POST)
 	public Result login(@RequestBody Admin admin){
+
 		Admin findAdmin = adminService.findByLoginName(admin.getLoginname(), admin.getPassword());
 		if (findAdmin == null || "".equals(findAdmin)){
 			return new Result(false, StatusCode.LOGINERROR, "用户名或密码错误");
 		}
-		return new Result(true,StatusCode.OK, "登录成功");
+		String token = jwtUtil.createJWT(findAdmin.getId(), findAdmin.getLoginname(), "admin");
+		Map<String, String> map = new HashMap<>();
+		map.put("loginname", findAdmin.getLoginname());
+		map.put("token", token);
+		return new Result(true,StatusCode.OK, "登录成功", map);
 	}
 	
 }
